+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+ - Parse trace state headers which contain a `=` character.
+
 ## 1.0.1
  - Align source and dist in published version.